SubmitProposalTxDetails component test
